rename event-emitter.ts -> mitt.ts
