+++ packages/next-server/lib/router/router.js
@@ -1,7 +1,7 @@
 /* global __NEXT_DATA__ */
 
 import { parse, format } from 'url'
-import mitt from '../event-emitter'
+import mitt from '../mitt'
 import shallowEquals from './shallow-equals'
 import { loadGetInitialProps, getURL } from '../utils'
