+++ packages/next/client/page-loader.js
@@ -1,5 +1,5 @@
 /* global document */
-import mitt from 'next-server/dist/lib/event-emitter'
+import mitt from 'next-server/dist/lib/mitt'
 
 // smaller version of https://gist.github.com/igrigorik/a02f2359f3bc50ca7a9c
 function supportsPreload (list) {
